wip: github: Migrate from Travis CI to Github Actions
On travis-ci.com we are now charged for each minute.  We only got 10000
credits in a trial plan, which we used up with a few builds.  Minutes also
cost a different amount of credits on different platforms: 10 on Linux,
but 50 on macOS (installing the dependencies on macOS alone took 12-15
minutes on Travis for some reason, works much faster on Github's runners).

No native Windows build yet as we have the same issue as on AppVeyor where
threading/streaming tests might get stuck.  And there is also only a
single platform to test on.  Plus building/testing on Windows is very
slow.

The 'sw_collector' test case had to be disabled because we can't access
/usr/local/share on the Github build hosts (the process is just blocked
in readdir() and eventually times out).

Unfortunately, we can't test on different architectures anymore (in
particular ARM and the big-endian IBM Z/x390x).